+++ client/src/pages/Dashboard.js
@@ -1,16 +1,16 @@
-// import React, { useState } from "react";
-import React from "react";
+import React, { useState } from "react";
+// import React from "react";
 import "../stylesheets/Dashboard.css";
 import { Link } from "react-router-dom";
-// import LoginForm from "./Forms/LoginForm";
+import LoginForm from "../components/Forms/LoginForm";
 
 export default function Dashboard() {
-  //   const { token, setToken } = useState();
+  const { token, setToken } = useState();
 
-  //   //THIS MAKES THE LOGIN PAGE THE FIRST THING ON THE WEBSIT
-  //   if (!token) {
-  //     return <LoginForm setToken={setToken} />;
-  //   }
+  //THIS MAKES THE LOGIN PAGE THE FIRST THING ON THE WEBSIT
+  if (!token) {
+    return <LoginForm setToken={setToken} />;
+  }
 
   return (
     <>
