+++ client/src/stylesheets/Form.css
@@ -7,7 +7,6 @@
 .content-wrap {
   min-height: 100vh;
   margin-bottom: -50px;
-  padding-bottom: 20px;
 }
 
 .formTitle {
@@ -71,6 +70,7 @@ input {
   font-size: 15px;
   margin-top: 20px;
   text-align: center;
+  padding-bottom: 20px;
 }
 
 /* .sign-up-sentence {
